# swagger.yaml
openapi: 3.0.2
info:
  title: To-do-list
  version: 1.0.0
paths:
  /users:
    summary: Path used to manage the list of users.
    description: >-
      The REST endpoint/path used to list and create zero or more `User` entities.  This path contains a
      `GET` and `POST` operation to perform the list and create tasks, respectively.
    post:
      requestBody:
        description: A new `User` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
            examples:
              User:
                value:
                  email: user@gmail.com
                  password: user
        required: true
      tags:
        - /users
      responses:
        '201':
          description: Successful response.
        '400':
          content:
            text/html:
              examples:
                Email and password are required:
                  value: Email and password are required
                'Unexpected token } in JSON at position 32':
                  value: 'Unexpected token } in JSON at position 32'
          description: Bad Request
        '409':
          content:
            text/html:
              examples:
                Email already exists:
                  value: Email already exists
          description: Email already exists
      operationId: createUser
      summary: Create a User
      description: Creates a new instance of a `User`.
  /sessions:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
            examples:
              Credentials:
                value:
                  email: user@gmail.com
                  password: user
        required: true
      tags:
        - /sessions
      responses:
        '201':
          content:
            application/json:
              examples:
                sessionId:
                  value:
                    sessionId: 8b359153-c7e4-4038-bbee-77638e9ee308
          description: 'Successful sign in '
        '400':
          content:
            text/html:
              examples:
                Email and password are required:
                  value: Email and password are required
          description: Bad Request
        '401':
          content:
            text/html:
              examples:
                Invalid email or password:
                  value: Invalid email or password
          description: Unauthorized
      summary: Sign in
    delete:
      tags:
        - /sessions
      responses:
        '204':
          description: Successful sign out
        '401':
          content:
            text/html:
              examples:
                Authorization header is required:
                  value: Authorization header is required
                Authorization header is invalid:
                  value: Authorization header is invalid
          description: Unauthorized
      security:
        - Bearer: [ ]
      summary: Sign out
components:
  schemas:
    User:
      title: Root Type for User
      description: ''
      type: object
      properties:
        id:
          format: int32
          type: integer
        email:
          type: string
        password:
          type: string
  securitySchemes:
    Bearer:
      scheme: bearer
      type: http
tags:
  - name: /users
    description: ''
  - name: /sessions
    description: ''